Socket

+++ game-start/src/main/java/com/lzh/game/start/util/PacketUtils.java
@@ -7,17 +7,17 @@ import com.lzh.game.start.model.player.service.SessionPlayerManage;
 
 public class PacketUtils {
 
-    public static void send(Player player, int cmd) {
+    public static void send(Player player, short cmd) {
         send(player, cmd, null);
     }
 
-    public static void send(Player player, int cmd, Object pack) {
+    public static void send(Player player, short cmd, Object pack) {
         SessionPlayerManage sessionManage = ApplicationUtils.getBean(SessionPlayerManage.class);
         Session session = sessionManage.getSessionByPlayer(player.getKey());
         send(session, cmd, pack);
     }
 
-    public static void send(Session session, int cmd, Object pack) {
+    public static void send(Session session, short cmd, Object pack) {
         session.write(SocketUtils.createOneWayRequest(cmd, pack));
     }
